# backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Swyngora Market API
  description: |
    Public market data for Swyngora (v0.1).

    - Candles, 24h volume/ticker, and spot market listing support **binance**, **coinbase**,
      and **bybit** (select with `?exchange=`).
    - Circulating / total / max supply is sourced from the **Binance** public marketing
      symbol list (asset-level, used for all venues' mcap enrichment), refreshed daily.
      Max supply is null when undefined.
    - Watchlist is unauthenticated state keyed by client-supplied `clientId`
      (must be non-empty; not for multi-tenant production without real auth).
      Data is persisted in SQLite on the server (survives restarts). Each list has a monotonic
      `version` for multi-device sync: send `baseVersion` on writes; concurrent adds of different
      symbols auto-merge; delete-vs-update on the same symbol returns 409 with both sides.
    - Price alerts support `one_time` (default) or `repeating` modes for above/below targets.
      One-time alerts become `triggered` after the first fire; repeating alerts stay `active`
      and re-fire only after the price returns to the safe side and crosses again.
    - Clients may register a webhook URL with deliveryMode `immediate` or `hourly_digest`,
      plus optional quiet hours (IANA timeZone + local start/end). Delivery is deferred until
      quiet hours end; pending notifications survive restarts.
...
      orders, and performance), market buy/sell at last price,
...
servers:
...
paths:
...
  /api/v1/portfolio/recurring-buys:
    post:
      summary: Create a paper recurring buy plan
      description: |
        Schedules named cash buys of a symbol at market (last) price.
        frequency: daily | weekly | monthly | interval.
        weekly + weekday (monday..sunday); monthly + dayOfMonth (1-31, salary day);
        interval + intervalHours (1-168, e.g. 12 = every 12 hours).
        Optional name (e.g. "Salary Day Buy"); default is "<symbol> <frequency>".
        Amount is cash notional spent each run. Insufficient cash fails that run only; the plan stays active.
        Missed periods execute only the latest due slot (no backlog of intermediate buys).
        Unique period keys prevent double execution across restarts or concurrent workers.
        Paper trading only — not real money.
      operationId: createRecurringBuyPlan
      tags: [Portfolio]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol, amount, frequency]
              properties:
                clientId: { type: string }
                exchange: { type: string, enum: [binance, coinbase, bybit], default: binance }
                symbol: { type: string, example: BTCUSDT }
                name:
                  type: string
                  maxLength: 80
                  description: User label e.g. Salary Day Buy; default is symbol + frequency
                  example: Salary Day Buy
                amount:
                  type: number
                  exclusiveMinimum: 0
                  description: Cash notional spent each run
                  example: 500
                frequency:
                  type: string
                  enum: [daily, weekly, monthly, interval]
                weekday:
                  type: string
                  enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
                  description: Weekly — buy on this UTC weekday
                dayOfMonth:
                  type: integer
                  minimum: 1
                  maximum: 31
                  description: Monthly — calendar day (clamped to last day of short months)
                intervalHours:
                  type: integer
                  minimum: 1
                  maximum: 168
                  description: Interval frequency — hours between buys (e.g. 12)
                startAt:
                  type: string
                  format: date-time
                  description: First scheduled run (RFC3339); default is now (next worker tick)
      responses:
        '201':
          description: Created plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringBuyPlan'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    get:
      summary: List paper recurring buy plans
      operationId: listRecurringBuyPlans
      tags: [Portfolio]
      parameters:
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Plan list
          content:
            application/json:
              schema:
                type: object
                properties:
                  clientId: { type: string }
                  plans:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecurringBuyPlan'
                  count: { type: integer }
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/recurring-buys/{id}:
    get:
      summary: Get a paper recurring buy plan
      operationId: getRecurringBuyPlan
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringBuyPlan'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      summary: Update a paper recurring buy plan
      description: Change name, amount, and/or schedule. Status still uses pause/resume.
      operationId: updateRecurringBuyPlan
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80 }
                amount: { type: number, exclusiveMinimum: 0 }
                frequency: { type: string, enum: [daily, weekly, monthly, interval] }
                weekday: { type: string, enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday] }
                dayOfMonth: { type: integer, minimum: 1, maximum: 31 }
                intervalHours: { type: integer, minimum: 1, maximum: 168 }
                startAt: { type: string, format: date-time }
      responses:
        '200':
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringBuyPlan'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      summary: Delete a paper recurring buy plan
      description: Removes the plan and its run history (cascade).
      operationId: deleteRecurringBuyPlan
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean }
                  id: { type: string }
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/recurring-buys/{id}/pause:
    post:
      summary: Pause a paper recurring buy plan
      description: Stops future executions; plan and history remain.
      operationId: pauseRecurringBuyPlan
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Paused plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringBuyPlan'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/recurring-buys/{id}/resume:
    post:
      summary: Resume a paused paper recurring buy plan
      description: Sets status active. If nextRunAt is in the past, it is bumped to now (next worker tick).
      operationId: resumeRecurringBuyPlan
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Active plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringBuyPlan'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/recurring-buys/{id}/runs:
    get:
      summary: List execution history for a recurring buy plan
      operationId: listRecurringBuyRuns
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Run list
          content:
            application/json:
              schema:
                type: object
                properties:
                  planId: { type: string }
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecurringBuyRun'
                  count: { type: integer }
        '404':
          $ref: '#/components/responses/Error'
...
  /api/v1/account:
    get:
      summary: Get account close/reopen status for a clientId
      operationId: getAccount
      tags: [Account]
      parameters:
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Account status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/account/close:
    post:
      summary: Close account (7-day grace period)
      description: |
        Marks the clientId closed. Product APIs return 403 for that clientId.
        Shared watchlists owned by this client become inaccessible to grantees.
        Data is retained until purgeAt (closedAt + 7 days). Active jobs are canceled.
      operationId: closeAccount
      tags: [Account]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                clientId: { type: string }
      responses:
        '200':
          description: Closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/account/reopen:
    post:
      summary: Reopen a closed account within the grace period
      operationId: reopenAccount
      tags: [Account]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                clientId: { type: string }
      responses:
        '200':
          description: Reopened (active)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Error'
...
components:
  responses:
    Error:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    RecurringBuyPlan:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string }
        exchange: { type: string }
        symbol: { type: string }
        name:
          type: string
          description: User label e.g. Salary Day Buy
        amount:
          type: number
          description: Cash notional spent each successful run
        frequency:
          type: string
          enum: [daily, weekly, monthly, interval]
        weekday:
          type: string
          enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        dayOfMonth:
          type: integer
          description: Monthly calendar day 1-31
        intervalHours:
          type: integer
          description: Hours between interval buys
        status:
          type: string
          enum: [active, paused]
        nextRunAt: { type: string, format: date-time }
        lastRunAt: { type: string, format: date-time, nullable: true }
        lastPeriodKey:
          type: string
          description: Last claimed period key (idempotency)
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RecurringBuyRun:
      type: object
      properties:
        id: { type: string }
        planId: { type: string }
        periodKey:
          type: string
          description: Calendar period key e.g. 2024-06-15 | 2024-W24 | 2024-06
        status:
          type: string
          enum: [succeeded, failed]
        amount: { type: number }
        quantity: { type: number }
        price: { type: number }
        tradeId: { type: string }
        failReason:
          type: string
          description: e.g. insufficient cash balance, market price unavailable
        scheduledFor: { type: string, format: date-time }
        executedAt: { type: string, format: date-time }
...
    Account:
      type: object
      properties:
        clientId: { type: string }
        status: { type: string, enum: [active, closed, purged] }
        closedAt: { type: string, format: date-time }
        purgeAt: { type: string, format: date-time }
        reopenedAt: { type: string, format: date-time }
        canReopen: { type: boolean }
        graceDays: { type: integer, example: 7 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...